Move message_conversion.hpp to src/io
